[fix] 업적은 엎어모달로.. #358
